[Windows] Support vulkan rhi build && run tests

+++ Code/Runtime/BerserkPlatform/BerserkPlatform/PlatformConsole.cpp
@@ -35,7 +35,7 @@ namespace Berserk {
     }
 
     void PlatformConsole::Write(uint64 symbolsCount, const String::Char8u *string) {
-        fwrite(string, sizeof(String::Char8u), symbolsCount, mOutputFile);
+        fwrite(string, sizeof(String::Char8u), static_cast<size_t>(symbolsCount), mOutputFile);
     }
 
     void PlatformConsole::Flush() {
